Add examples for depth, preorder and post order
