+++ src/examples/tree-traversal/preorder.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-import { accept, createFromString } from '../index';
+import { accept, createFromString } from '../../index';
 import ts from 'typescript';
 
 // Use https://ts-ast-viewer.com to check the AST.
@@ -10,11 +10,8 @@ export function f(): void;
 
 const ast = createFromString(declarationFile);
 
-const depthVisitor = {
+accept(ast, {
   pre: (node: ts.Node) => {
-    console.log(`${node.pos}:${node.end}`);
+    console.log(`${node.pos}:${node.end}:${node.kind}`);
   },
-};
-
-console.log(`Pre order:`);
-accept(ast, depthVisitor);
+});
