+++ src/getChildrenWithoutTokens.ts
@@ -1,0 +1,11 @@
+import ts from 'typescript';
+
+export const getChildrenWithoutTokens = (node: ts.Node) => {
+  const children: ts.Node[] = [];
+
+  ts.forEachChild(node, (child) => {
+    children.push(child);
+  });
+
+  return children;
+};
